requirements.txt
fastai==2.7.12
pandas==1.3.4
Pillow==8.4.0
Pillow==9.5.0
